+++ ichnaea/api/locate/tests/test_fallback.py
@@ -24,8 +24,9 @@ class TestSource(BaseSourceTest):
     TestSource = FallbackPositionSource
     settings = {
         'url': 'http://127.0.0.1:9/?api',
-        'ratelimit': '10',
+        'ratelimit': '3',
         'ratelimit_expire': '60',
+        'ratelimit_interval': '2',
         'cache_expire': '60',
     }